Fixed library games appending to each other.
http://i.imgur.com/Vu9u2Tf.png
The library element gets reset before the async function, so if someone
makes more than one request in a short period of time (before the
previous request ends), the informations are merged together, resulting
in the anomaly shown on the screenshot.
This code makes sure that the downloaded data is for the currently
selected game on the list, and runs the "creation" procedure only if it
is. In case the request wasn't successful, it shows the usual "App not
found" message, and in case the request was successful, but for another
game, it basically gets discarded. The loading indicator stays until the
proper game data gets downloaded, or in case it already was, nothing
happens.

what did i write idonteven