Use gulp to JSHint the project.
Added a very simple gulpfile to JSHint the project, this ignores jQuery
as we know this is already defined.

The `.jshintrc` contains linting rules, these can be changed.

`.travis-yml` enables CI builds like this:
https://travis-ci.org/Saturate/sharepoint-forms-ext, when the projects
uses UNIT Tests it will be easy to make it test these as well.

To install gulp please look at their website. You need to run `npm
install` to install the dependencies defined in package.json

Longer down the road it would be a good idea to have a source folder,
and a build folder that could be packaged for chrome ect. this way the
project is more clean, especcialy if you are minizming the files via
gulp.

Questions / comments?