Fixes increased row height on listings page
Using "margin-top: 13px;" on the cloud image to align it vertically increases the height of the main row, multiply this by a few hundred times (based on how many listing you have) and the overall page height is increased quite a bit. To fix it use "vertical-align: middle;" instead.
See https://imgur.com/OQRKXcv for comparison.