Multiple requestes due to fn call from "each"
"each" runs async. which causes hundreds of requestes(while viewing the
wishlist at least) based on how items are on page, to
store.steampowered.com/dynamicstore/userdata/ when calling
highlight_notinterested() if the "dynamiclist_time" has expired or not
set.
This causes the browser be very slow or might even freeze on slower
computers for a few seconds.

This is a fast fix since async. requeste are deprecated so we should
look into fixing this properly.